Using a secondary constructor instead of inheritance
